backend/requirements.txt
Django==4.2.30
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
Pillow==12.2.0
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.1
django-filter==23.5
psycopg2-binary==2.9.9
gunicorn==23.0.0
whitenoise==6.6.0
